--- a/packages/sequencer/src/protocol/production/sequencing/BlockProductionService.ts
+++ b/packages/sequencer/src/protocol/production/sequencing/BlockProductionService.ts
@@ -104,13 +104,12 @@ export class BlockProductionService {
 
     const lastResult = lastBlockWithResult.result;
     const lastBlock = lastBlockWithResult.block;
-    const executionResults: TransactionExecutionResult[] = [];
 
     const incomingMessagesList = new MinaActionsHashList(
       Field(lastBlock.toMessagesHash)
     );
 
-    let blockState: BlockTrackers = {
+    const blockState: BlockTrackers = {
       blockHashRoot: Field(lastResult.blockHashRoot),
       eternalTransactionsList: new TransactionHashList(
         lastBlock.toEternalTransactionsHash
@@ -133,28 +132,13 @@ export class BlockProductionService {
       UntypedStateTransition.fromStateTransition(transition)
     );
 
-    for (const tx of transactions) {
-      try {
-        // Create execution trace
-        const [newState, executionTrace] =
-          // eslint-disable-next-line no-await-in-loop
-          await this.transactionExecutionService.createExecutionTrace(
-            stateService,
-            tx,
-            networkState,
-            blockState
-          );
-
-        blockState = newState;
-
-        // Push result to results and transaction onto bundle-hash
-        executionResults.push(executionTrace);
-      } catch (error) {
-        if (error instanceof Error) {
-          log.error("Error in inclusion of tx, skipping", error);
-        }
-      }
-    }
+    const [newBlockState, executionResults] =
+      await this.transactionExecutionService.createExecutionTraces(
+        stateService,
+        transactions,
+        networkState,
+        blockState
+      );
 
     const previousBlockHash =
       lastResult.blockHash === 0n ? undefined : Field(lastResult.blockHash);
@@ -168,9 +152,10 @@ export class BlockProductionService {
 
     const block: Omit<Block, "hash"> = {
       transactions: executionResults,
-      transactionsHash: blockState.transactionList.commitment,
+      transactionsHash: newBlockState.transactionList.commitment,
       fromEternalTransactionsHash: lastBlock.toEternalTransactionsHash,
-      toEternalTransactionsHash: blockState.eternalTransactionsList.commitment,
+      toEternalTransactionsHash:
+        newBlockState.eternalTransactionsList.commitment,
       height:
         lastBlock.hash.toBigInt() !== 0n ? lastBlock.height.add(1) : Field(0),
       fromBlockHashRoot: Field(lastResult.blockHashRoot),