Based on #292 

- Improved block.commit from 660ms to 67ms
	- Memoized PendingTransaction hash -> saved around 370ms (for 100 txs)
	- Removed unnecessary "create if missing" for transactions in a block, since they already exist in the table because of the mempool using the same table
- Reduce block.transaction overhead by 1ms or so by creating a NonProvable toStateTransitionHash implementation (saves us a few hashes)

Block production with 100 txs with unique keys and postgres/redis DB config:
Before: 5357ms
Now: 2994ms (32.82 tps)